% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso.R
\name{lasso}
\alias{lasso}
\title{Fit a Lasso Path}
\usage{
lasso(
  x,
  y,
  lambda = NULL,
  path_length = 100,
  lambda_min_ratio = if (NROW(x) > NCOL(x)) 1e-04 else 0.01,
  tol = 1e-06,
  max_it = 10000,
  warm_starts = TRUE
)
}
\arguments{
\item{x}{Design matrix}

\item{y}{Response vector}

\item{lambda}{Lambda sequence}

\item{tol}{Stopping tolerance in terms of the duality gap relative to the
value of the primal objective for the null model}

\item{max_it}{Number of maximum iterations for the solver}
}
\value{
family Type of objective
}
\description{
Fit a Lasso Path
}
