% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_path.R
\name{lasso_path}
\alias{lasso_path}
\title{Lasso Path with Hessian Screening Rules}
\usage{
lasso_path(
  x,
  y,
  family = c("gaussian", "binomial"),
  lambda = NULL,
  standardize = TRUE,
  shuffle = FALSE,
  check_frequency = if (NROW(x) > NCOL(x)) 1 else 10,
  hessian_warm_starts = TRUE,
  augment_with_gap_safe = TRUE,
  log_hessian_update_type = c("full", "approx"),
  log_hessian_auto_update_freq = 10,
  path_length = 100L,
  maxit = 1e+05,
  tol_gap = 1e-04,
  gamma = 0.01,
  store_dual_variables = FALSE,
  verify_hessian = FALSE,
  line_search = TRUE,
  verbosity = 0
)
}
\arguments{
\item{y}{The reponse vector}

\item{family}{The name of the family, "gaussian" or "logistic"}

\item{lambda}{weights for the regularization path, if \code{NULL}, then they
are automatically computed}

\item{standardize}{Whether to standardize the predictors}

\item{shuffle}{Shuffle working set before each CD pass?}

\item{check_frequency}{Frequency at which duality gap is checked for
inner CD loop}

\item{hessian_warm_starts}{Whether to use warm starts based on Hessian}

\item{augment_with_gap_safe}{Whether or not augment heuristic rules
with Gap Safe checks during KKT checks}

\item{log_hessian_update_type}{What type of strategy to use for
updating the hessian for logistic regression}

\item{path_length}{The (desired) length of the lasso path}

\item{maxit}{Maximum number of iterations for Coordinate Descent loop}

\item{tol_gap}{Tolerance threshold for relative duality gap.}

\item{gamma}{Percent of strong approximation to add to Hessian approximation}

\item{store_dual_variables}{whether to store dual variables throughout
fitting}

\item{verify_hessian}{Whether to not to verify that Hessian updates are
correct. Used only for diagnostic purposes.}

\item{line_search}{Use line search in CD solver.}

\item{verbosity}{Controls the level of verbosity. 0 = no output, 1 = outer
level output, 2 = inner solver output}

\item{X}{The predictor matrix}
}
\description{
This function fits the full lasso path.
}
