% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hesso-package.R
\docType{package}
\name{hesso-package}
\alias{hesso}
\alias{hesso-package}
\title{hesso: Lasso Paths with Hessian Screening and Warm Starts}
\description{
Procedures for fitting L1-regularized least-squares regression (the lasso) using Hessian screening and warm starts.
}
\author{
\strong{Maintainer}: Johan Larsson \email{johan.larsson@stat.lu.se} (\href{https://orcid.org/0000-0002-4029-5945}{ORCID})

Authors:
\itemize{
  \item Jonas Wallin \email{jonas.wallin@stat.lu.se} (\href{https://orcid.org/0000-0003-0381-6593}{ORCID})
}

}
\keyword{internal}
