% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso.R
\name{lasso}
\alias{lasso}
\title{Fit a Lasso Path}
\usage{
lasso(
  x,
  y,
  intercept = TRUE,
  lambda = NULL,
  path_length = 100,
  lambda_min_ratio = if (NROW(x) > NCOL(x)) 1e-04 else 0.01,
  tol = 1e-06,
  max_it = 10000,
  warm_starts = TRUE
)
}
\arguments{
\item{x}{Design matrix. If sparse needs to be of class \code{dgCMatrix}. If dense,
needs to be a matrix (i.e. not a \code{data.frame}).}

\item{y}{Response vector}

\item{intercept}{Whether to fit an intercept}

\item{lambda}{Lambda sequence}

\item{path_length}{The length of the path. Ignored if \code{lambda} is not \code{NULL}.}

\item{lambda_min_ratio}{A fraction of the largest lambda in the path, which
is the last point on the path. Ignored if \code{lambda} is not NULL.}

\item{tol}{Stopping tolerance in terms of the duality gap relative to the
value of the primal objective for the null model}

\item{max_it}{Number of maximum iterations for the solver}
}
\value{
family Type of objective
}
\description{
Fit a Lasso Path
}
